+++ arraypageradapter/build.gradle
@@ -84,6 +84,8 @@ install {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test:runner:1.1.0-beta02'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-beta02'
     implementation 'androidx.appcompat:appcompat:1.0.0'
 }
 task sourcesJar(type: Jar) {
