+++ demo/src/main/java/com/demo/arraypageradapter/ViewPagerAdapterActivity.java
@@ -1,6 +1,4 @@
-package com.demo.arraypargeradapter;
-
-
+package com.demo.arraypageradapter;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +6,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.demo.arraypargeradapter.R;
 import com.nakama.arraypageradapter.ArrayViewPagerAdapter;
 
 import androidx.appcompat.app.AppCompatActivity;
