+++ demo/src/main/res/layout/activity_pager.xml
@@ -4,10 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin">
+    android:paddingBottom="@dimen/activity_vertical_margin">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/view_pager"
@@ -15,7 +15,7 @@
         android:layout_height="@dimen/view_pager_height"
         android:background="@color/colorBackground" />
 
-    <com.demo.arraypargeradapter.ControlView
+    <com.demo.arraypageradapter.ControlView
         android:id="@+id/control_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
