+++ demo/src/main/res/layout/widget_control.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<ScrollView
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:android="http://schemas.android.com/apk/res/android">
+    android:layout_height="match_parent">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -14,8 +13,7 @@
             android:id="@+id/add_btn"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Add">
-        </Button>
+            android:text="Add" />
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -34,7 +32,7 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="2"
-                android:hint="AddPosition"
+                android:hint="Position"
                 android:inputType="number"
                 android:text="0" />
         </LinearLayout>
@@ -56,21 +54,21 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="2"
-                android:hint="RemovePosition"
+                android:hint="Position"
                 android:inputType="number"
                 android:text="0" />
         </LinearLayout>
 
         <Button
             android:id="@+id/add_all_btn"
-            android:text="Add 3 Items"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            android:text="Add 3 Items" />
 
         <Button
             android:id="@+id/clear_btn"
-            android:text="Clear"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            android:text="Clear" />
     </LinearLayout>
 </ScrollView>
