+++ demo/src/main/res/values/dimens.xml
@@ -3,5 +3,4 @@
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="view_pager_height">200dp</dimen>
-
 </resources>
