+++ demo/src/test/java/com/demo/arraypageradapter/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.nakama.arrayviewpager;
+package com.demo.arraypageradapter;
 
 import org.junit.Test;
